Fix failure in yang-maven-plugin-it
- With the introduction of the odlparent, the IT mentioned in the
subject start to fails in autorelease because the location of the root
pom file is another repository, attached via the settings.xml file.
Now org.apache.maven.it.Verifier when is called in that IT doesn't
have a proper settings.xml, and unfortunately due to:

https://jira.codehaus.org/browse/SUREFIRE-790

the failsafe and surefire plugins don't get the location of the user
settings propagated down, hence odlparent could not be found by the
IT.
The workaround is to pre-populate the maven.repo.local with the
artifacts released in the previous cycle, so there is actually no need
for Verifier to go and hunt for the artifacts outside. The class
org.apache.maven.it.Verifier at least honor the maven.repo.local,
which is also passed down by surefire else the testcases would simply
leave sandbox environment.

Signed-off-by: Giovanni Meo <gmeo@cisco.com>