Allow the case where patches to a branch cannot be applied
- Usually when there is a problem on the release of a project, a patch
can be cached and then applied before doing the release. Now if that
patch is upstreamed in the branch the release process will stop
working because the patch will not apply. Now it's perfectly ok to
actually try to abort and keep trying, maybe it's the good case and
build will succeed

Signed-off-by: Giovanni Meo <gmeo@cisco.com>