Enhacend wagon-maven-plugin to support maven release repositories
- Maven release repository don't allow the redeployment of the
artifacts but allow to redeploy the metadata to support the addition
of new artifacts. In a case of a failure while uploading a repository
that maybe is transient it's beneficial to allow a new run of the
pushrelease script without worrying that existing artifacts will fail
to redeploy. This enhancement for all the artifacts that are already
in the repository skips the upload, however will retry to upload the
metadata. This way if there is a failure in the transfer it's a
genuine error and not due to a redeploy of a released artifact that
anyway we know in a release repository is normal situation

Signed-off-by: Giovanni Meo <gmeo@cisco.com>